componentize home view
