+++ src/components/currency/TargetView.vue
@@ -1,13 +1,15 @@
 <script>
+  import { mapGetters } from "vuex";
+
   export default {
-    props: ["currencySymbol", "value"]
+    computed: mapGetters(["targetCurrencySymbol", "targetValue"])
   };
 </script>
 
 
 <template>
   <v-layout>
-    <span class="headline">{{ this.currencySymbol }}</span>
-    <h3 class="display-2">{{ this.value }}</h3>
+    <span class="headline">{{ this.targetCurrencySymbol }}</span>
+    <h3 class="display-2">{{ this.targetValue }}</h3>
   </v-layout>
 </template>
